Added training part to bi- , trigram and tinynn. Almost finished checkpoints.

+++ src/trainer/Options.cs
@@ -21,5 +21,5 @@ public class Options
     public int Seed { get; set; }
 
     [Option("lr", Default = 0.1, HelpText = "Learning rate (для TinyNN)")]
-    public double LearningRate { get; set; }
+    public float LearningRate { get; set; }
 }
